+++ src/main/java/persistence/bootstrap/binder/CollectionLoaderBinder.java
@@ -1,8 +1,8 @@
-package persistence.bootstrap;
+package persistence.bootstrap.binder;
 
 import jdbc.JdbcTemplate;
-import jdbc.RowMapper;
-import persistence.entity.CollectionLoader;
+import jdbc.mapper.RowMapper;
+import persistence.entity.loader.CollectionLoader;
 import persistence.meta.EntityTable;
 import persistence.sql.dml.DmlQueries;
