+++ src/main/java/persistence/bootstrap/binder/EntityTableBinder.java
@@ -1,4 +1,4 @@
-package persistence.bootstrap;
+package persistence.bootstrap.binder;
 
 import persistence.meta.EntityTable;
