+++ src/main/java/persistence/entity/manager/EntityEntry.java
@@ -1,4 +1,4 @@
-package persistence.entity;
+package persistence.entity.manager;
 
 import java.util.List;
 import java.util.Objects;
