+++ src/main/java/persistence/entity/manager/EntityManager.java
@@ -1,4 +1,4 @@
-package persistence.entity;
+package persistence.entity.manager;
 
 public interface EntityManager {
     <T> T find(Class<T> clazz, Object id);
