+++ src/test/java/persistence/entity/EntityLoaderTest.java
@@ -4,12 +4,15 @@ import database.H2ConnectionFactory;
 import domain.Order;
 import domain.OrderItem;
 import fixture.EntityWithId;
-import jdbc.DefaultRowMapper;
 import jdbc.JdbcTemplate;
+import jdbc.mapper.DefaultRowMapper;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import persistence.entity.loader.CollectionLoader;
+import persistence.entity.loader.EntityLoader;
+import persistence.entity.manager.EntityManager;
 import persistence.entity.proxy.ProxyFactory;
 import persistence.meta.EntityTable;
 import persistence.sql.dml.SelectQuery;
