+++ src/test/java/persistence/entity/PersistenceContextTest.java
@@ -3,6 +3,9 @@ package persistence.entity;
 import fixture.EntityWithId;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import persistence.entity.manager.EntityEntry;
+import persistence.entity.manager.EntityStatus;
+import persistence.entity.manager.factory.PersistenceContext;
 import persistence.meta.EntityKey;
 import persistence.meta.EntityTable;
