+++ src/test/java/persistence/entity/proxy/ProxyFactoryTest.java
@@ -3,15 +3,14 @@ package persistence.entity.proxy;
 import database.H2ConnectionFactory;
 import domain.OrderItem;
 import domain.OrderLazy;
-import jdbc.DefaultRowMapper;
 import jdbc.JdbcTemplate;
+import jdbc.mapper.DefaultRowMapper;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import persistence.entity.CollectionLoader;
-import persistence.entity.EntityManager;
-import persistence.entity.LazyLoader;
+import persistence.entity.loader.CollectionLoader;
+import persistence.entity.manager.EntityManager;
 import persistence.meta.EntityTable;
 import persistence.sql.dml.SelectQuery;
 import util.TestHelper;
